+++ lib/objectbox/actions.dart
@@ -8,6 +8,7 @@ import "package:flow/data/money.dart";
 import "package:flow/data/multi_currency_flow.dart";
 import "package:flow/data/prefs/frecency_group.dart";
 import "package:flow/data/transaction_filter.dart";
+import "package:flow/data/transaction_programmable_object.dart";
 import "package:flow/data/transactions_filter/time_range.dart";
 import "package:flow/entity/account.dart";
 import "package:flow/entity/backup_entry.dart";
@@ -23,6 +24,8 @@ import "package:flow/l10n/extensions.dart";
 import "package:flow/objectbox.dart";
 import "package:flow/objectbox/objectbox.g.dart";
 import "package:flow/prefs/local_preferences.dart";
+import "package:flow/services/accounts.dart";
+import "package:flow/services/categories.dart";
 import "package:flow/services/exchange_rates.dart";
 import "package:flow/services/file_attachment.dart";
 import "package:flow/services/recurring_transactions.dart";
@@ -970,17 +973,20 @@ extension AccountActions on Account {
     TransactionSubtype? subtype,
     Recurrence? recurrence,
     List<String>? extraTags,
-    List<double>? location,
+    double? latitude,
+    double? longitude,
   }) {
     FileAttachmentService().upsertManySync(attachments ?? []);
 
     final String uuid = uuidOverride ?? const Uuid().v4();
 
-    if (location == null) {
+    if (latitude == null || longitude == null) {
       final Geo? geo =
           extensions?.firstWhereOrNull((ext) => ext is Geo) as Geo?;
 
-      location ??= geo?.toLatLng();
+      final latLng = geo?.toLatLng();
+      latitude ??= latLng?[0];
+      longitude ??= latLng?[1];
     }
 
     final String? recurringTransactionUuid = recurrence == null
@@ -1014,6 +1020,9 @@ extension AccountActions on Account {
             isPending: isPending ?? false,
             subtype: subtype?.value,
             extraTags: extraTags ?? [],
+            location: (latitude != null && longitude != null)
+                ? [latitude, longitude]
+                : null,
           )
           ..setTags(tags ?? [])
           ..setCategory(category)
@@ -1126,6 +1135,106 @@ extension BackupEntryActions on BackupEntry {
       return ObjectBox().box<BackupEntry>().remove(id);
     } catch (e) {
       return false;
+    }
+  }
+}
+
+extension TransactionProgrammableObjectActions
+    on TransactionProgrammableObject {
+  int save(dynamic fromAccount, {dynamic toAccount}) {
+    Account? targetAccount = switch (fromAccount) {
+      Account a => a,
+      dynamic any => AccountsService().findOneActiveSync(any),
+    };
+
+    targetAccount ??= AccountsService().findOneActiveSync(
+      UserPreferencesService().primaryAccountUuid,
+    );
+
+    Account? targetTransferToAccount = switch (toAccount) {
+      Account a => a,
+      dynamic any => AccountsService().findOneActiveSync(any),
+    };
+
+    if (targetAccount == null) {
+      throw Exception(
+        "Failed to find account for TransactionProgrammableObject.save",
+      );
+    }
+
+    if (type == .transfer &&
+        targetTransferToAccount == null &&
+        targetAccount.uuid != UserPreferencesService().primaryAccountUuid) {
+      targetTransferToAccount = AccountsService().findOneActiveSync(
+        UserPreferencesService().primaryAccountUuid,
+      );
+    }
+
+    if (type == .transfer && targetTransferToAccount == null) {
+      throw Exception(
+        "Failed to find target account for transfer TransactionProgrammableObject.save",
+      );
+    }
+
+    Category? resolvedCategory;
+
+    if (type != .transfer) {
+      try {
+        resolvedCategory =
+            CategoriesService().findOneSync(categoryUuid) ??
+            CategoriesService().findOneSync(category);
+      } catch (e) {
+        // Ignore errors here
+      }
+    }
+
+    if (type == .transfer) {
+      return targetAccount
+          .transferTo(
+            targetAccount: targetTransferToAccount!,
+            amount: amount ?? 0.0,
+            title: title,
+            description: notes,
+            transactionDate: transactionDate,
+            isPending: isPending,
+
+            // tags: tags,
+            // attachments: attachments,
+            // conversionRate: conversionRate,
+            // recurrence: recurrence,
+            // extraTags: extraTags,
+          )
+          .$1;
+    } else {
+      // required double amount,
+      // DateTime? transactionDate,
+      // DateTime? createdDate,
+      // String? title,
+      // String? description,
+      // Category? category,
+      // List<TransactionExtension>? extensions,
+      // List<TransactionTag>? tags,
+      // List<FileAttachment>? attachments,
+      // String? uuidOverride,
+      // bool? isPending,
+      // TransactionSubtype? subtype,
+      // Recurrence? recurrence,
+      // List<String>? extraTags,
+      // List<double>? location,
+      return targetAccount.createAndSaveTransaction(
+        amount: amount ?? 0.0,
+        transactionDate: transactionDate,
+        title: title,
+        description: notes,
+        category: resolvedCategory,
+        // tags: tags,
+        // attachments: attachments,
+        isPending: isPending,
+        // subtype: subtype,
+        // extraTags: extraTags,
+        latitude: lat,
+        longitude: lng,
+      );
     }
   }
 }